fix: forgot to close scenarios
