+++ app/src/androidTest/java/com/github/ybecker/epforuml/EditPhotoTest.kt
@@ -13,6 +13,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.dsphotoeditor.sdk.activity.DsPhotoEditorActivity
 import org.junit.Test
+import org.junit.rules.ExpectedException
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
@@ -40,5 +41,24 @@ class EditPhotoTest {
         }
         Intents.release()
         scenario.close()
+    }
+
+    @Test
+    fun failsWithIncorrectUri(){
+        var exception = ExpectedException.none()
+        val intent = Intent(
+            ApplicationProvider.getApplicationContext(),
+            EditPhotoActivity::class.java
+        )
+
+        intent.putExtra("uri", "INVALIDURI")
+
+        val scenario = ActivityScenario.launch<Activity>(intent)
+        exception.expect(IllegalArgumentException::class.java)
+
+
+        Intents.release()
+        scenario.close()
+
     }
 }
